# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Alphabet AI - Adaptive K-12 ELA Mastery Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List ELA skills with optional filters
      parameters:
        - in: query
          name: gradeLevel
          schema:
            type: string
        - in: query
          name: domain
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: active
          schema:
            type: boolean
      responses:
        "200":
          description: List of skills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ElaSkill"
  /skills/{skillCode}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get a single skill by code
      parameters:
        - in: path
          name: skillCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ElaSkill"
        "404":
          description: Not found
  /skills/tree:
    get:
      operationId: getSkillTree
      tags: [skills]
      summary: Get skills grouped by grade and domain for skill tree view
      parameters:
        - in: query
          name: gradeLevel
          schema:
            type: string
      responses:
        "200":
          description: Skill tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillTree"

  # ── MASTERY ────────────────────────────────────────────────────────────────
  /mastery:
    get:
      operationId: listMastery
      tags: [mastery]
      summary: List skill mastery records for current student
      parameters:
        - in: query
          name: domain
          schema:
            type: string
        - in: query
          name: masteryLevel
          schema:
            type: string
      responses:
        "200":
          description: Mastery records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillMastery"
  /mastery/{skillCode}:
    get:
      operationId: getMastery
      tags: [mastery]
      summary: Get mastery for a specific skill
      parameters:
        - in: path
          name: skillCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mastery record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillMastery"
        "404":
          description: Not found
  /mastery/{skillCode}/record:
    post:
      operationId: recordAnswer
      tags: [mastery]
      summary: Record a practice answer and update mastery
      parameters:
        - in: path
          name: skillCode
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerRecord"
      responses:
        "200":
          description: Updated mastery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillMastery"
...
  /llm/generate-question:
    post:
      operationId: generateQuestion
      tags: [llm]
      summary: Generate an adaptive practice question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionGenerateInput"
      responses:
        "200":
          description: Generated question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdaptiveQuestion"
...
components:
  schemas:
...
    ElaSkill:
      type: object
      required: [skillCode, skillName, gradeLevel, domainCode]
      properties:
        skillCode:
          type: string
        skillName:
          type: string
        description:
          type: ["string", "null"]
        gradeLevel:
          type: string
        gradeBand:
          type: ["string", "null"]
        domainCode:
          type: string
        domain:
          type: string
        substrand:
          type: ["string", "null"]
        standardLeafCode:
          type: ["string", "null"]
        parentGseCode:
          type: ["string", "null"]
        parentCcssCode:
          type: ["string", "null"]
        difficulty:
          type: number
        discrimination:
          type: number
        guessing:
          type: number
        subSkillOrder:
          type: integer
        prerequisiteSkillCodes:
          type: array
          items:
            type: string
        nextSkillCodes:
          type: array
          items:
            type: string
        culturallyRelevantThemes:
          type: array
          items:
            type: string
        sorPillar:
          type: ["string", "null"]
        active:
          type: boolean
    SkillTree:
      type: object
      required: [gradeLevel, domains]
      properties:
        gradeLevel:
          type: string
        domains:
          type: array
          items:
            $ref: "#/components/schemas/SkillDomain"
    SkillDomain:
      type: object
      required: [domainCode, domain, skills]
      properties:
        domainCode:
          type: string
        domain:
          type: string
        colorHex:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/ElaSkill"

    # ── MASTERY ───────────────────────────────────────────────────────────────
    SkillMastery:
      type: object
      required: [id, studentId, skillCode, skillName, masteryLevel, smartScore, practiceCount]
      properties:
        id:
          type: string
        studentId:
          type: string
        skillCode:
          type: string
        skillName:
          type: string
        domain:
          type: ["string", "null"]
        masteryLevel:
          type: string
          enum: [not_started, introduced, practicing, approaching, mastered]
        masteryPercentage:
          type: number
        smartScore:
          type: number
        practiceCount:
          type: integer
        correctCount:
          type: integer
        needsReteaching:
          type: boolean
        isUnlocked:
          type: boolean
        lastPracticed:
          type: ["string", "null"]
    AnswerRecord:
      type: object
      required: [correct, questionDifficulty]
      properties:
        correct:
          type: boolean
        questionDifficulty:
          type: number
        timeSpentSeconds:
          type: integer
...
    AdaptiveQuestion:
      type: object
      required: [id, skillCode, skillName, questionText, activityType, options]
      properties:
        id:
          type: string
        skillCode:
          type: string
        skillName:
          type: string
        domain:
          type: string
        questionText:
          type: string
        passage:
          type: ["string", "null"]
        activityType:
          type: string
          enum: [listen_repeat, see_tap, say_it, write_it, read_it, multiple_choice]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        correctOptionId:
          type: string
        explanation:
          type: ["string", "null"]
        difficulty:
          type: number
    QuestionOption:
      type: object
      required: [id, text]
      properties:
        id:
          type: string
        text:
          type: string
...
    QuestionGenerateInput:
      type: object
      required: [skillCode, gradeLevel, activityType]
      properties:
        skillCode:
          type: string
        gradeLevel:
          type: string
        activityType:
          type: string
        interests:
          type: array
          items:
            type: string
        culturalContext:
          type: array
          items:
            type: string
        homeLanguage:
          type: string
        theta:
          type: number
...